Entities and controllers all updated for real data

+++ api/src/main/java/com/example/api/dto/ReplyDto.java
@@ -3,10 +3,10 @@ package com.example.api.dto;
 import java.time.LocalDateTime;
 
 
-public record ReplyDto (int id, 
-                      int bulletin_id,
+public record ReplyDto (Long id, 
+                      Long bulletin_id,
                       LocalDateTime posted_on, 
-                      int posted_by, 
+                      Long posted_by, 
                       String content) {
   
 }
